Split interaction_test
